* Elimicación form para generar consulta * Modificación de variables * Cambio en algunas estructuras de Main
