+++ COMPILADORES/Main.cs
@@ -16,18 +16,50 @@ namespace COMPILADORES
     {
         string instruccionCorrecta = "";
         int tipo = 0;
+        String CCadena = "";
+        SqlConnection conn;
+        analizador.AnalisisSintactico analizer = new analizador.AnalisisSintactico();
+        String conexion;
+
         public Main()
         {
             InitializeComponent();
         }
+        public void ConsultarDatos(SqlCommand cmd)
+        {
+            SqlDataReader reader = cmd.ExecuteReader();
+            Resultado resultado = new Resultado(conexion, txtSQL.Text);
+            resultado.Show();
+        }
+        public void LlenarComboBox(){
+            System.Data.SqlClient.SqlConnection SqlCon = new System.Data.SqlClient.SqlConnection(CCadena);
+            SqlCon.Open();
 
-        SqlConnection conn;
+            System.Data.SqlClient.SqlCommand SqlCom = new System.Data.SqlClient.SqlCommand();
+            SqlCom.Connection = SqlCon;
+            SqlCom.CommandType = CommandType.StoredProcedure;
+            SqlCom.CommandText = "sp_databases";
 
-        analizador.AnalisisSintactico ana = new analizador.AnalisisSintactico();
+            System.Data.SqlClient.SqlDataReader SqlDR;
+            SqlDR = SqlCom.ExecuteReader();
 
-        //String de conexion
-        String conexion;
-        private void button1_Click(object sender, EventArgs e)
+            List<String> listBD = new List<String>();
+
+            while (SqlDR.Read())
+            {
+                comboBox1.Items.Add(SqlDR.GetValue(0));
+            }
+        }
+
+        private void Main_Load(object sender, EventArgs e)
+        {
+            this.Size = new Size(803, 527);
+            groupBox2.Visible = false;
+            Ejecucion.Enabled = false;
+            txtResultado.ReadOnly = true;
+            comboBox1.DropDownStyle = ComboBoxStyle.DropDownList;
+        }
+        private void Ejecición_Click(object sender, EventArgs e)
         {
             if (txtSQL.Text == "")
             {
@@ -91,79 +123,7 @@ namespace COMPILADORES
             }
 
         }
-
-        private void mostrarErrores(int op, ref ArrayList errores)
-        {
-            if (op == 1) // Si son errores lexicos
-            {
-                foreach (string s in errores)
-                    txtResultado.Text = "La palabra: '" + s + "' no pertenece al lenguaje del parserSQL. \r\n";
-
-            }
-            else // Si son errores sintacticos
-            {
-                foreach (string s in errores)
-                txtResultado.Text += s + "\r\n";
-            }
-        }
-
-        public void ConsultarDatos(SqlCommand cmd)
-        {
-            SqlDataReader reader = cmd.ExecuteReader();
-            Resultado resultado = new Resultado(conexion, txtSQL.Text);
-            resultado.Show();
-        }
-
-        private void button2_Click(object sender, EventArgs e)
-        {
-            string campo  ="";
-            ana.analisis(txtSQL.Text, ref campo);
-            //creando conexion a sql
-            conn = new SqlConnection();
-            try
-            {
-                conn.ConnectionString = txtStringCon.Text;
-                conn.Open();
-                llenarComboBox();
-                conn.Close();
-
-                MessageBox.Show("Prueba de conexión exitosa", "Informe", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            }
-            catch (Exception)
-            {
-                MessageBox.Show("Prueba de conexión fallida", "Informe", MessageBoxButtons.OK, MessageBoxIcon.Error);
-            }
-            
-        }
-
-        public void llenarComboBox(){
-            System.Data.SqlClient.SqlConnection SqlCon = new System.Data.SqlClient.SqlConnection(txtStringCon.Text);
-            SqlCon.Open();
-
-            System.Data.SqlClient.SqlCommand SqlCom = new System.Data.SqlClient.SqlCommand();
-            SqlCom.Connection = SqlCon;
-            SqlCom.CommandType = CommandType.StoredProcedure;
-            SqlCom.CommandText = "sp_databases";
-
-            System.Data.SqlClient.SqlDataReader SqlDR;
-            SqlDR = SqlCom.ExecuteReader();
-
-            List<String> listBD = new List<String>();
-
-            while (SqlDR.Read())
-            {
-                comboBox1.Items.Add(SqlDR.GetValue(0));
-            }
-        }
-
-        private void Main_Load(object sender, EventArgs e)
-        {
-            button1.Enabled = false;
-            txtResultado.ReadOnly = true;
-            comboBox1.DropDownStyle = ComboBoxStyle.DropDownList;
-        }
-
-        private void button2_Click_1(object sender, EventArgs e)
+        private void btnGenerarConsulta_Click_1(object sender, EventArgs e)
         {
             if (comboBox1.Items.Count <= 0)
             {
@@ -175,25 +135,122 @@ namespace COMPILADORES
             }
             else
             {
-                Consulta consulta = new Consulta();
-                consulta.Show();
+                this.Size = new Size(803 + 423, 527);
+                groupBox2.Visible = true;
             }
         }
 
-        private void button3_Click(object sender, EventArgs e)
-        {
-            BD ventanaBD = new BD(txtStringCon.Text);
-            ventanaBD.Show();
-        }
+
 
         private void comboBox1_SelectedIndexChanged(object sender, EventArgs e)
         {
-            string[] part = txtStringCon.Text.Split(';');
+            string[] part = CCadena.Split(';');
             conexion = part[0] + ";Initial Catalog=" + comboBox1.SelectedItem.ToString() + ";" + part[1];
-            button1.Enabled = true;
-            button3.Enabled = true;
-            button2.Enabled = true;
+            Ejecucion.Enabled = true;
+            DBCatalog.Enabled = true;
+            btnGenerarConsulta.Enabled = true;
             //MessageBox.Show(part[0]+";Initial Catalog="+comboBox1.SelectedItem.ToString()+";"+part[1]);
+        }
+
+        private void btnGenerar_Click(object sender, EventArgs e)
+        {
+            if (txtWhere.Text != "")
+            {
+                if (txtOrderBy.Text != "")
+                {
+                    txtConsulta.Text = txtOriginal.Text + " WHERE " + txtWhere.Text + " ORDER BY " + txtOrderBy.Text;
+                    txtSQL.Text = txtConsulta.Text;
+                }
+                else
+                {
+                    txtConsulta.Text = txtOriginal.Text + " WHERE " + txtWhere.Text;
+                    txtSQL.Text = txtConsulta.Text;
+                }
+            }
+            else
+            {
+                if (txtOrderBy.Text != "")
+                {
+                    txtConsulta.Text = txtOriginal.Text + " ORDER BY " + txtOrderBy.Text;
+                    txtSQL.Text = txtConsulta.Text;
+                }
+                else
+                {
+                    txtConsulta.Text = txtOriginal.Text;
+                    txtSQL.Text = txtConsulta.Text;
+                }
+            }
+        }
+
+        private void Cancelar_Click(object sender, EventArgs e)
+        {
+            this.Size = new Size(803, 527);
+            groupBox2.Visible = false;
+            txtConsulta.Text = "";
+            txtOrderBy.Text = "";
+            txtOriginal.Text = "";
+            txtResultado.Text = "";
+        }
+
+        private void btnconn_Click(object sender, EventArgs e)
+        {
+            string campo = "";
+            analizer.analisis(txtSQL.Text, ref campo);
+            //creando conexion a sql
+            conn = new SqlConnection();
+            try
+            {
+                if (((txtUser.Text == "" || txtUser.Text == null) && (txtPassword.Text == "" || txtPassword.Text == null)) || (txtUser.Text == "" || txtUser.Text == null) || (txtPassword.Text == "" || txtPassword.Text == null))
+                {
+                    conn.ConnectionString = "Data Source=" + txtSorce.Text + ";Integrated Security=True";
+                }
+                else
+                {
+                    conn.ConnectionString = "Data Source=" + txtSorce.Text + ";User Id=" + txtUser.Text + ";Password=" + txtPassword.Text;
+                }
+                CCadena = conn.ConnectionString;
+                try
+                {
+                    conn.Open();
+                }
+                catch
+                {
+                    MessageBox.Show(":/");
+                }
+                LlenarComboBox();
+                conn.Close();
+
+                MessageBox.Show("Prueba de conexión exitosa", "Informe", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+            catch (Exception)
+            {
+                MessageBox.Show("Prueba de conexión fallida", "Informe", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+
+        }
+
+        private void DBCatalog_Click_1(object sender, EventArgs e)
+        {
+            Estructura ventanaBD = new Estructura(CCadena);
+            ventanaBD.Show();
+        }
+        private void mostrarErrores(int op, ref ArrayList errores)
+        {
+            if (op == 1) // Si son errores lexicos
+            {
+                foreach (string s in errores)
+                {
+                    txtResultado.Text = "La palabra: '" + s + "' no pertenece al lenguaje del parserSQL. \r\n";
+                }
+
+            }
+            else // Si son errores sintacticos
+            {
+                foreach (string s in errores)
+                {
+                    txtResultado.Text += s + "\r\n";
+                }
+            }
         }
     }
 }
